clean up some duplicated code
