+++ src/main/scala/kamon/newrelic/metrics/NewRelicDistributionMetrics.scala
@@ -10,7 +10,7 @@ import com.newrelic.telemetry.metrics.{Gauge, Metric, Summary}
 import kamon.metric.Distribution.Bucket
 import kamon.metric.MetricSnapshot.Distributions
 import kamon.metric.{Distribution, DynamicRange, MeasurementUnit, MetricSnapshot}
-import kamon.newrelic.TagsToAttributes
+import kamon.newrelic.AttributeBuddy
 import kamon.tag.TagSet
 import org.slf4j.LoggerFactory
 
@@ -25,7 +25,7 @@ object NewRelicDistributionMetrics {
 
     dist.instruments.flatMap { inst =>
       val tags: TagSet = inst.tags
-      val instrumentBaseAttributes: Attributes = TagsToAttributes.addTagsFromTagSets(Seq(tags), baseAttributes.copy())
+      val instrumentBaseAttributes: Attributes = AttributeBuddy.addTagsFromTagSets(Seq(tags), baseAttributes.copy())
 
       val distValue: Distribution = inst.value
