+++ src/main/scala/kamon/newrelic/metrics/NewRelicMetricsReporter.scala
@@ -5,13 +5,13 @@
 
 package kamon.newrelic.metrics
 
+import com.newrelic.telemetry.SimpleMetricBatchSender
 import com.newrelic.telemetry.metrics.{MetricBatch, MetricBatchSender}
-import com.newrelic.telemetry.{Attributes, SimpleMetricBatchSender}
 import com.typesafe.config.Config
 import kamon.Kamon
 import kamon.metric.PeriodSnapshot
 import kamon.module.{MetricReporter, Module, ModuleFactory}
-import kamon.newrelic.TagsToAttributes
+import kamon.newrelic.AttributeBuddy.buildCommonAttributes
 import org.slf4j.LoggerFactory
 
 import scala.jdk.CollectionConverters._
@@ -21,21 +21,6 @@ class NewRelicMetricsReporter(senderBuilder: () => MetricBatchSender = () => New
   private val logger = LoggerFactory.getLogger(classOf[NewRelicMetricsReporter])
   @volatile private var commonAttributes = buildCommonAttributes(Kamon.config())
   @volatile private var sender: MetricBatchSender = senderBuilder()
-
-  private def buildCommonAttributes(config: Config) = {
-    val environment = config.getConfig("kamon.environment")
-    val serviceName = if (environment.hasPath("service")) environment.getString("service") else null
-    val host = if (environment.hasPath("host")) environment.getString("host") else null
-    val attributes = new Attributes()
-      .put("instrumentation.source", "kamon-agent")
-      .put("service.name", serviceName)
-      .put("host", host)
-    if (environment.hasPath("tags")) {
-      val environmentTags = environment.getConfig("tags")
-      TagsToAttributes.addTagsFromConfig(environmentTags, attributes)
-    }
-    attributes
-  }
 
   override def reportPeriodSnapshot(snapshot: PeriodSnapshot) = {
     logger.warn("NewRelicMetricReporter reportPeriodSnapshot...")
